requirements.txt
Django>=3.0.4,<3.1.0
djangorestframework>=3.11.0,<3.12.0
django-cors-headers>=3.2.1,<3.3.0
psycopg2>=2.8.4,<2.9.0
flake8>=3.7.9,<3.8.0
isort>=5.10.0,<5.20.0
djangorestframework-simplejwt>=4.6.0,<4.7.0
PyJWT>=2.0.1,<2.1.0
pdfplumber==0.11.5
openai==1.60.2
PyYAML>=5.4.1,<5.5.0
coreapi>=2.3.3,<2.4.0